// src/models/Proveedor.ts
// Generated by https://quicktype.io

export interface Proveedor {
  id: number;
  nombre: string;
  rfc: string;
  calle: string;
  colonia: string;
  telefono: string;
  ciudad: string;
  estado: string;
  cp: string;
  contacto: string;
  email: string;
  observaciones: string;
  nombre_fiscal: string;
  dias_financiamiento: number;
  fecha_alta: string;
  fecha_act: string;
}